Update the models to use partial classes
