+++ UID.Participant.Api/Models/Participant.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using UID.Participant.Api.Validation;
 
 namespace UID.Participant.Api.Models;
 
@@ -16,6 +15,5 @@ public partial class Participant
 
     public bool Visible { get; set; }
 
-    [ValidateClientTypes]
     public virtual ICollection<ClientType> ClientTypes { get; set; } = new List<ClientType>();
 }
